+++ src/model/IPath.ts
@@ -1,0 +1,4 @@
+export interface IPath {
+  id: string;
+  value: string;
+}